requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
altair==6.0.0
    # via streamlit
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==7.0.5
    # via streamlit
certifi==2026.2.25
    # via requests
charset-normalizer==3.4.7
    # via requests
click==8.3.2
    # via streamlit
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
idna==3.11
    # via requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
narwhals==2.19.0
    # via
    #   altair
    #   plotly
numpy==2.4.4
    # via
    #   -r requirements.in
    #   pandas
    #   pydeck
    #   scipy
    #   streamlit
oauthlib==3.3.1
    # via requests-oauthlib
packaging==26.0
    # via
    #   altair
    #   plotly
    #   streamlit
pandas==3.0.2
    # via
    #   -r requirements.in
    #   streamlit
pillow==12.2.0
    # via streamlit
plotly==6.6.0
    # via -r requirements.in
protobuf==7.34.1
    # via streamlit
pyarrow==23.0.1
    # via
    #   -r requirements.in
    #   streamlit
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via pandas
pyyaml==6.0.3
    # via -r requirements.in
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.33.1
    # via
    #   requests-oauthlib
    #   streamlit
requests-oauthlib==2.0.0
    # via -r requirements.in
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
scipy==1.17.1
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
smmap==5.0.3
    # via gitdb
streamlit==1.56.0
    # via -r requirements.in
tenacity==9.1.4
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.5
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   streamlit
urllib3==2.6.3
    # via requests
